+++ src/helpers/parse-keyword-value.ts
@@ -12,6 +12,7 @@
 export const parseKeywordValue = (rawValue: string): string => {
     const NOW_VALUE_KEYWORD = '$now$';
     const CURRENT_DATE_KEYWORD = '$currentDate$';
+    const CURRENT_ISO_DATE_KEYWORD = '$currentISODate$';
 
     let parsedValue = rawValue;
 
@@ -21,6 +22,9 @@ export const parseKeywordValue = (rawValue: string): string => {
     } else if (rawValue === CURRENT_DATE_KEYWORD) {
         // Set to current date e.g 'Tue Nov 08 2022 13:53:19 GMT+0300'
         parsedValue = Date();
+    } else if (rawValue === CURRENT_ISO_DATE_KEYWORD) {
+        // Set to current date e.g '2022-11-08T13:53:19.650Z'
+        parsedValue = new Date().toISOString();
     }
 
     return parsedValue;